requirements.txt
requests~=2.32.3
uvicorn~=0.30.6
fastapi~=0.112.1
pydantic~=2.8.2
python-dotenv~=1.0.1
openai~=1.41.0